requirements.txt
streamlit>=1.35.0
streamlit-folium>=0.20.0
folium>=0.16.0
plotly>=5.20.0
pandas>=2.0.0
numpy>=1.26.0
scikit-learn>=1.4.0
xgboost>=2.0.0
google-generativeai>=0.5.0
geopy>=2.4.0
requests>=2.31.0
Pillow>=10.0.0
reportlab>=4.0.0
